+++ codegen/src/main/java/ch/loway/oss/ari4java/codegen/genJava/JavaGen.java
@@ -18,44 +18,37 @@ public class JavaGen {
                 .append("//               PLEASE DO NOT EDIT                    \n")
                 .append("//    Generated on: ").append((new Date()).toString()).append("\n")
                 .append("// ----------------------------------------------------\n\n");
-
     }
 
     public static void importClasses(StringBuilder sb, String myPackage, List<String> imports) {
-
         addPackage(sb, myPackage);
         for (String pkg : imports) {
             sb.append("import ").append(pkg).append(";\n");
         }
         sb.append("\n");
-
     }
 
     public static void addBanner(StringBuilder sb, String multiLineBanner) {
-
         String[] rows = multiLineBanner.split("\n");
         sb.append("/**\n");
         for (String row : rows) {
             if (!row.isEmpty()) {
                 row = row.replaceAll("<br /><br />", "\n * ");
                 row = row.replaceAll("<br />", "\n * ");
+                row = row.replaceAll("<", "&gt;").replaceAll(">", "&lt;");
                 sb.append(" * ").append(row).append("\n");
             }
         }
         sb.append(" */\n");
-
     }
 
     public static void addBanner(StringBuilder sb, String multilineBanner, String sinceVersion) {
-        multilineBanner += "\n@since " + sinceVersion;
+        multilineBanner += "\n@since " + sinceVersion.toUpperCase();
         addBanner(sb, multilineBanner);
     }
 
     public static String addPrefixAndCapitalize(String prefix, String field) {
-
         return prefix + field.substring(0, 1).toUpperCase() + field.substring(1);
-
-
     }
 
     public static String addAsyncCallback(String response) {
@@ -68,12 +61,4 @@ public class JavaGen {
         }
     }
 
-    public static void emptyLine(StringBuilder sb) {
-        emptyLines(sb, 1);
-    }
-
-
 }
-
-// $Log$
-//