+++ codegen/src/main/java/ch/loway/oss/ari4java/codegen/run.java
@@ -2,7 +2,6 @@
 package ch.loway.oss.ari4java.codegen;
 
 import java.io.File;
-import java.io.IOException;
 import java.nio.file.Files;
 import java.nio.file.Paths;
 
@@ -13,17 +12,16 @@ import java.nio.file.Paths;
  */
 public class run {
 
-    private static String SOURCES = "./src/main/resources/codegen-data/";
-    private static String OUTPUT = "../src/main/generated/";
-
-    public static void main(String[] argv) throws IOException {
+    public static void main(String[] argv) throws Exception {
         System.out.println("This is ARI4JAVA Code Generator version " + VERSION.VER);
+        String sourceFolder = "./src/main/resources/codegen-data/";
+        String outputFolder = "../src/main/generated/";
 
         DefMapper dm = new DefMapper();
-        dm.setOutputFolder(OUTPUT);
+        dm.setOutputFolder(outputFolder);
         dm.clean();
 
-        Files.list(Paths.get(SOURCES))
+        Files.list(Paths.get(sourceFolder))
                 .filter(p -> p.toFile().isDirectory())
                 .sorted((p1, p2) -> {
                     // break up folder into version parts and sort as numbers
@@ -43,17 +41,17 @@ public class run {
                 .forEach(p -> loadAsteriskDefs(dm, p.toFile()));
 
         dm.generateAllClasses();
-//        System.out.println("Finished Generating");
-
     }
 
     private static void loadAsteriskDefs(DefMapper dm, File folder) {
         try {
             String srcVer = folder.getName();
-//            System.out.println("Loading: " + folder.getAbsolutePath());
-            for (File definition : folder.listFiles()) {
-                if (definition.getName().endsWith(".json")) {
-                    dm.parseJsonDefinition(definition, srcVer);
+            File[] files = folder.listFiles();
+            if (files != null) {
+                for (File definition : files) {
+                    if (definition.getName().endsWith(".json")) {
+                        dm.parseJsonDefinition(definition, srcVer);
+                    }
                 }
             }
         } catch (Exception e) {
@@ -63,5 +61,3 @@ public class run {
 
 }
 
-// $Log$
-//